requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==4.2
django-rest-framework==0.1.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyYAML==6.0.1
python-dotenv~=1.0.1
